[#318] Spruce up test harness
This commit makes some updates to the files defining the test harness
so that it will be better-suited to run in a GitHub Actions Workflow.
There are other things that need to happen in order to support iRODS 5
and generally improve the situation, but this is a minimal effort to
enable automated testing.